# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DataForSEO Intelligence API
  version: 1.0.0
  summary: Multi-tenant SEO research API on top of DataForSEO v3.
  description: 'A multi-tenant SaaS SEO platform. Every billed call flows through a three-tier cost engine
    (Redis -> Postgres -> upstream) with singleflight coalescing and stale-while-revalidate, so repeat
    reads are $0.


    **Conventions**

    - All endpoints are versioned under `/api/v1`.

    - Every data response embeds a `meta` object (`from_cache`, `cost_cents`, `source`, `latency_ms`).

    - All money values are integer **USD cents**.

    - Auth is JWT Bearer; access tokens are short-lived, refreshed via `/auth/refresh`.

    - Per-organization monthly quota is enforced before any billed call (HTTP 402).'
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Proprietary
paths:
...
  /api/v1/projects:
    get:
      tags:
      - projects
      summary: List Projects
      operationId: list_projects_api_v1_projects_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Opaque pagination cursor.
          title: Cursor
        description: Opaque pagination cursor.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ProjectOut_'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - projects
      summary: Create Project
      operationId: create_project_api_v1_projects_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectOut'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/projects/{project_id}:
    get:
      tags:
      - projects
      summary: Get Project
      operationId: get_project_api_v1_projects__project_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
      - projects
      summary: Update Project
      operationId: update_project_api_v1_projects__project_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectOut'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - projects
      summary: Delete Project
      operationId: delete_project_api_v1_projects__project_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '204':
          description: Successful Response
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/projects/{project_id}/runs:
    post:
      tags:
      - projects
      summary: Save Run
      operationId: save_run_api_v1_projects__project_id__runs_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRunCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRunOut'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags:
      - projects
      summary: List Runs
      operationId: list_runs_api_v1_projects__project_id__runs_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Opaque pagination cursor.
          title: Cursor
        description: Opaque pagination cursor.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ProjectRunOut_'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/projects/{project_id}/runs/{run_id}:
    get:
      tags:
      - projects
      summary: Open Run
      operationId: open_run_api_v1_projects__project_id__runs__run_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRunResult'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    CursorPage:
      properties:
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
        has_more:
          type: boolean
          title: Has More
          default: false
      type: object
      title: CursorPage
...
    Page_ProjectOut_:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ProjectOut'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/CursorPage'
      type: object
      required:
      - data
      - pagination
      title: Page[ProjectOut]
    Page_ProjectRunOut_:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ProjectRunOut'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/CursorPage'
      type: object
      required:
      - data
      - pagination
      title: Page[ProjectRunOut]
    ProjectCreate:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        type:
          type: string
          title: Type
          default: keyword
        config:
          type: object
          title: Config
      type: object
      required:
      - name
      title: ProjectCreate
    ProjectDetail:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        type:
          type: string
          title: Type
        config:
          type: object
          title: Config
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        run_count:
          type: integer
          title: Run Count
          default: 0
        runs:
          items:
            $ref: '#/components/schemas/ProjectRunOut'
          type: array
          title: Runs
      type: object
      required:
      - id
      - name
      - type
      - config
      - created_at
      - updated_at
      title: ProjectDetail
    ProjectOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        type:
          type: string
          title: Type
        config:
          type: object
          title: Config
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        run_count:
          type: integer
          title: Run Count
          default: 0
      type: object
      required:
      - id
      - name
      - type
      - config
      - created_at
      - updated_at
      title: ProjectOut
    ProjectRunCreate:
      properties:
        module:
          type: string
          maxLength: 50
          minLength: 1
          title: Module
        params:
          type: object
          title: Params
        result:
          type: object
          title: Result
      type: object
      required:
      - module
      title: ProjectRunCreate
    ProjectRunOut:
      properties:
        id:
          type: string
          title: Id
        module:
          type: string
          title: Module
        params:
          type: object
          title: Params
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - module
      - params
      - created_at
      title: ProjectRunOut
    ProjectRunResult:
      properties:
        id:
          type: string
          title: Id
        module:
          type: string
          title: Module
        params:
          type: object
          title: Params
        result:
          type: object
          title: Result
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - module
      - params
      - result
      - created_at
      title: ProjectRunResult
    ProjectUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Name
        config:
          anyOf:
          - type: object
          - type: 'null'
          title: Config
      type: object
      title: ProjectUpdate
...
    Problem:
      type: object
      description: RFC 7807 Problem Details (media type `application/problem+json`).
      required:
      - type
      - title
      - status
      - detail
      properties:
        type:
          type: string
          format: uri
          examples:
          - https://docs.seo-intelligence.app/errors/quota-exceeded
        title:
          type: string
          examples:
          - Payment Required
        status:
          type: integer
          examples:
          - 402
        detail:
          type: string
          examples:
          - Monthly organization quota exhausted.
        instance:
          type: string
          examples:
          - /api/v1/serp/ranking
        errors:
          type: array
          description: Field-level validation failures (422 only).
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from `/auth/login` or `/auth/refresh`.
  responses:
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: The requested resource does not exist or is not visible to you.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationFailed:
      description: Request body or parameters failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.example.com
  description: Production
...
security:
- BearerAuth: []